+++ resources/core/cljs/client.cljs
@@ -8,6 +8,7 @@
             <</hydrogen-session?>>[day8.re-frame.http-fx]
             [re-frame.core :as rf]
             [reagent.dom :as rd]
+            [<<namespace>>.client.breadcrumbs :as breadcrumbs]
             [<<namespace>>.client.home :as home]<<#hydrogen-session?>>
             [<<namespace>>.client.landing :as landing]<</hydrogen-session?>>
             [<<namespace>>.client.routes :as routes]<<#hydrogen-session-keycloak?>>
@@ -53,6 +54,7 @@
         :class (str "theme-" (name @theme))}
        [:div.app-container__main
         {:id "app-container__main"}
+        [breadcrumbs/main]
         (case @active-view
           :foo [:div "FIXME"]
           [home/main])]